+++ packages/yourloops/lib/error/error.api.ts
@@ -24,7 +24,7 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-import HttpService from '../../services/http.service'
+import HttpService from '../http/http.service'
 
 export interface ErrorPayload {
   browserName: string