+++ packages/yourloops/test/unit/lib/notifications/notification.api.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2022, Diabeloop
+ * Copyright (c) 2021-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -26,7 +26,7 @@
  */
 
 import HttpStatus from '../../../../lib/http/models/enums/http-status.enum'
-import HttpService, { ErrorMessageStatus } from '../../../../services/http.service'
+import HttpService, { ErrorMessageStatus } from '../../../../lib/http/http.service'
 import { loggedInUsers } from '../../common'
 import axios, { AxiosResponse } from 'axios'
 import NotificationApi from '../../../../lib/notifications/notification.api'